feat: Added ability to change background based on form fields
As per the initial instructions, I set the background image equal to the
Google Streetview image of the address submitted in the form fields.

In addition, I made it so the image takes up the appropriate amount of space
in the browser window by using the window dimensions to size the background-size
css attribute of the body.